ClojureScript: self-host arity fix.
Apparently ClojureScript functions that have metadata attached only
support arity of 20. This is problem during self-host because the main
EVAL cond macro is 22 items so it blows up when applied in
macroexpand. So we preserve the origin unadorned function through to
macroexpand and use it there instead of the adorned function/macro.